Rename and organazie db codes better
